requirements.txt
meshtastic~=2.7.3
Pypubsub~=4.0.3